+++ tools/build-pydroid.ps1
@@ -143,7 +143,7 @@ param(
     [int]$PnpmNetworkConcurrency = 16
 )
 
-$script:BuildScriptRevision = "1.6.51-dev-r146-group-primitive-node-polish"
+$script:BuildScriptRevision = "1.6.52-dev-r147-primitive-micro-ui-polish"
 
 $ErrorActionPreference = "Stop"
 try { [Console]::OutputEncoding = New-Object System.Text.UTF8Encoding($false) } catch {}